UPSTREAM: drm/i915/dsb: Evade transcoder undelayed vblank when using DSB
We want to start the DSB execution from the transcoder's undelayed
vblank, so in order to guarantee atomicity with the all the other
mmio register writes we need to evade both vblanks.

Note that currently we don't add any vblank delay, so this is
effectively a nop. But in the future when we start to program
double buffered registers from the DSB we'll need to delay the
pipe's vblank to provide the register programming "window2"
for the DSB.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230606191504.18099-15-ville.syrjala@linux.intel.com
Reviewed-by: Uma Shankar <uma.shankar@intel.com>
(cherry picked from commit 7678e089bd18b4dd61478ce728f01fd5239e97ff)

BUG=b:325713868
TEST=tast run <IP> graphics.IgtKms.kms_vrr

Change-Id: I379861d462f1116730b2fb3f20ed39716c0a77ea
Signed-off-by: Tim Van Patten <timvp@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5429980
Reviewed-by: Manasi Navare <navaremanasi@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>