UPSTREAM: regulator: mt6358: Use mt6397-regulator.h binding header for buck mode macros
The (undocumented) possible values for the buck operating modes on the
MT6358 are the same as those on the MT6397, both for the device tree
bindings and the actual hardware register values.

Reuse the macros for the MT6397 PMIC in the MT6358 regulator driver by
including the mt6397-regulator.h binding header and replacing the
existing macros. This aligns it with other PMIC.

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Link: https://lore.kernel.org/r/20230928085537.3246669-7-wenst@chromium.org
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 9f3bec54d06f1eb4b47c7f78ef1401bc71977e9e)

BUG=b:362203280
TEST=emerge-corsola-kernelnext chromeos-kernel-6_6

Change-Id: I7d494dcecf8d2e2cbd7acdfb5f365b86f2e52f55
Signed-off-by: Albert Jakieła <jakiela@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5817827
Reviewed-by: Wojciech Macek <wmacek@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Wojciech Macek <wmacek@google.com>