UPSTREAM: arm64: dts: mediatek: mt8195-cherry: Assign sram supply to MFG1 pd
Add a phandle to the MT8195_POWER_DOMAIN_MFG1 power domain and
assign the GPU SRAM (vsram_others) supply to that in mt8195-cherry:
this allows to keep the sram powered up while the GPU is used.

This means that it's now possible to remove the regulator-always-on
property from the mt6359_vsram_others_ldo_reg vreg, so that it will
be switched on and off during suspend.

Tested-by: Chen-Yu Tsai <wenst@chromium.org>
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
(cherry picked from commit 3106b14c1cb4b745dd0413dc392418d301f3b1d1)

BUG=b:330841611
TEST=boot Cherry and confirm graphics and OpenGL are OK

Change-Id: I19ba895dde5b7e80e73586870e557c1ca242f77b
Signed-off-by: Wojciech Macek <wmacek@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5387737
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>