CHROMIUM: iwl7000: take UEFI lock status out of ACPI ifdef
iwl_fw_runtime::uefi_tables_lock_status is a value read from ACPI
and is indicating whether the UEFI tables are locked.
The driver checks this value before deciding if to read a UEFI table or
not.

The lock status is currently defined under ifdef ACPI,
but the code checking it is not, and shouldn't be. So non-ACPI
compilations will fail.

Take it out of ACPI ifdef, and then for non-ACPI compilation this will
be false, as required.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I77e74a35bebba4466462cda4273b57837b7fc656
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: d99e2b29309626195497f6b7a6f58bc5104c3af2