UPSTREAM: wifi: cfg80211: detect stuck ECSA element in probe resp
We recently added some validation that we don't try to
connect to an AP that is currently in a channel switch
process, since that might want the channel to be quiet
or we might not be able to connect in time to hear the
switching in a beacon. This was in commit c09c4f31998b
("wifi: mac80211: don't connect to an AP while it's in
a CSA process").

However, we promptly got a report that this caused new
connection failures, and it turns out that the AP that
we now cannot connect to is permanently advertising an
extended channel switch announcement, even with quiet.
The AP in question was an Asus RT-AC53, with firmware
3.0.0.4.380_10760-g21a5898.

As a first step, attempt to detect that we're dealing
with such a situation, so mac80211 can use this later.

Reported-by: coldolt <andypalmadi@gmail.com>
Closes: https://lore.kernel.org/linux-wireless/CAJvGw+DQhBk_mHXeu6RTOds5iramMW2FbMB01VbKRA4YbHHDTA@mail.gmail.com/
Fixes: c09c4f31998b ("wifi: mac80211: don't connect to an AP while it's in a CSA process")
Reviewed-by: Miriam Rachel Korenblit <miriam.rachel.korenblit@intel.com>
Link: https://msgid.link/20240129131413.246972c8775e.Ibf834d7f52f9951a353b6872383da710a7358338@changeid
Change-Id: Iaa1d061585cd4fdb57e6954dc6086c6025ad6e7d
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 177fbbcb4ed6b306c1626a277fac3fb1c495a4c7)

BUG=b:355129747
TEST=emerge-brya chromeos-kernel-6_6

Change-Id: I86d49bd055582051878ad10c62d3741dab45fc6d
Signed-off-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5758620
Reviewed-by: David Ruth <druth@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>