UPSTREAM: arm64: dts: mediatek: mt8183-pico6: Fix wake-on-X event node names
The wake-on-bt and wake-on-wlan nodes don't have a button- or event-
prefix that the gpio-keys binding requires.

Fix up the node names to satisfy the binding. While at it, also fix up
the GPIO overriding structure for the wake-on-wlan node. Instead of
referencing the gpio-keys node and then open coding the node, add a
label for the event node, and use that to reference and override the
GPIO settings.

Fixes: 055ef10ccdd4 ("arm64: dts: mt8183: Add jacuzzi pico/pico6 board")
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Link: https://lore.kernel.org/r/20240131084043.3970576-1-wenst@chromium.org
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
(cherry picked from commit 70bf81dd2c2dedbed1f19dfc5d1d2f22474a5296)

BUG=b:360780931
TEST=Build kernel

Change-Id: Ia5349a1359961597d10a567419afcfce06dacb9b
Signed-off-by: Hsin-Te Yuan <yuanhsinte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5799426
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>